Creating layered filter component
