+++ src/app/components/map/map.component.html
@@ -18,9 +18,6 @@
                     [displayControls]="false">
 </app-visible-layers>
 
-<app-filter [(displayFilter)]="displayFilter">
-</app-filter>
-
 <app-legend [(selectedLayers)]="selectedLayers"
             [(displayLegend)]="displayLegend">
 </app-legend>
