+++ src/app/components/sidebar/sidebar-menu/sidebar-item/sidebar-item.component.html
@@ -20,7 +20,7 @@
         </ng-template>
 
     </div>
-
+    
     <div class="sidebar-item-switch">
 
         <p-inputSwitch *ngIf="!sidebarItem.link"
