+++ src/app/components/sidebar/sidebar.component.html
@@ -7,4 +7,8 @@
                     [sidebarItems]="sidebarItems">
   </app-sidebar-menu>
 
+
+  <app-filter [(displayFilter)]="displayFilter"></app-filter>
+
+
   <app-sidebar-footer class="sidebar-footer"></app-sidebar-footer>
