+++ src/app/components/sidebar/sidebar.component.ts
@@ -1,23 +1,29 @@
-import { Component, OnInit } from '@angular/core';
+import {AfterViewInit, Component, Input, OnInit} from '@angular/core';
 
 import { ConfigService } from '../../services/config.service';
 
 import { LayerGroup } from 'src/app/models/layer-group.model';
 
 import { Layer } from 'src/app/models/layer.model';
+import * as L from 'leaflet';
 
 @Component({
   selector: 'app-sidebar',
   templateUrl: './sidebar.component.html',
   styleUrls: ['./sidebar.component.css']
 })
-export class SidebarComponent implements OnInit {
+export class SidebarComponent implements OnInit, AfterViewInit {
+
+  @Input() displayFilterControl = true;
+
   sidebarItems: LayerGroup[] = [];
 
   sidebarConfig;
 
   logoPath: string;
   logoLink: string;
+
+  displayFilter = false;
 
   constructor(
     private configService: ConfigService
@@ -28,6 +34,34 @@ export class SidebarComponent implements OnInit {
     this.logoPath = this.sidebarConfig.logoPath;
     this.logoLink = this.sidebarConfig.logoLink;
     this.setSidebarItems();
+  }
+
+  ngAfterViewInit() {
+    if (this.displayFilterControl) {
+      this.setFilterControl();
+    }
+  }
+
+  setFilterControl() {
+    const Filter = L.Control.extend({
+      onAdd: () => {
+        const div = L.DomUtil.create('div');
+        div.innerHTML = `
+          <div id="filterBtn" class="leaflet-control-layers leaflet-custom-icon" title="Filtro">
+            <a><i class='fas fa-filter'></i></a>
+          </div>`;
+        return div;
+      }
+    });
+    //
+    // new Filter({ position: 'topleft' }).addTo(this.map);
+
+    this.setFilterControlEvent();
+  }
+
+  setFilterControlEvent() {
+    L.DomEvent.on(L.DomUtil.get('filterBtn'), 'dblclick', L.DomEvent.stopPropagation);
+    document.querySelector('#filterBtn').addEventListener('click', () => this.displayFilter = !this.displayFilter);
   }
 
   setSidebarItems() {
@@ -64,6 +98,10 @@ export class SidebarComponent implements OnInit {
       layerGroup.children = layerChildren;
       this.sidebarItems.push(layerGroup);
     });
+  }
+
+  filterClick(event) {
+    console.log(event);
   }
 
 }
